# swagger_ui/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "DZ API.",
    "version": "0.0.1"
  },
  "host": "localhost:8181",
  "basePath": "/api/",
  "paths": {
...
    "/stocks": {
      "get": {
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Stocks"
        ],
        "summary": "Get all Stocks",
        "operationId": "getAllStocks",
        "responses": {
          "200": {
            "$ref": "#/responses/stocks"
          }
        }
      },
      "post": {
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Stocks"
        ],
        "summary": "Create a Stock",
        "operationId": "createStock",
        "parameters": [
          {
            "name": "Stock",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Stock"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/stock"
          }
        }
      }
    },
    "/stocks/{id}": {
      "put": {
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Stocks"
        ],
        "summary": "Update a Stock by id",
        "operationId": "updateStockById",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/stock"
          }
        }
      },
      "delete": {
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Stocks"
        ],
        "summary": "Delete a Stock by id",
        "operationId": "deleteStockById",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/stock"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Lists all users.",
        "operationId": "getUsers",
        "responses": {
          "200": {
            "$ref": "#/responses/users"
          },
          "401": {
            "$ref": "#/responses/notAuthorized"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Consumes;\napplication/json",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get User by id.",
        "operationId": "getUserById",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/user"
          },
          "401": {
            "$ref": "#/responses/notAuthorized"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Branch": {
      "description": "Branch is a model bro",
      "type": "object",
      "properties": {
        "address": {
          "type": "string",
          "x-go-name": "Address"
        },
        "branch": {
          "type": "string",
          "x-go-name": "Branch"
        },
        "date_created": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateCreated"
        },
        "date_updated": {
          "description": "the dateUpdated for the product",
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateUpdated"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID",
          "readOnly": true
        }
      },
      "x-go-package": "github.com/aiman-zaki/go_dz_http/models"
    },
...
    "ShiftWork": {
      "type": "object",
      "properties": {
        "date_created": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateCreated"
        },
        "date_updated": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateUpdated"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "shift": {
          "type": "string",
          "x-go-name": "Shift"
        }
      },
      "x-go-package": "github.com/aiman-zaki/go_dz_http/models"
    },
    "Stock": {
      "description": "Stock : model",
      "type": "object",
      "properties": {
        "branch": {
          "$ref": "#/definitions/Branch"
        },
        "branch_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "BranchID"
        },
        "date_created": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateCreated"
        },
        "date_updated": {
          "description": "the dateUpdated for the product",
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateUpdated"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID",
          "readOnly": true
        },
        "shift_work": {
          "$ref": "#/definitions/ShiftWork"
        },
        "shift_work_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ShiftWorkID"
        },
        "stock_date": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "StockDate"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "user_id": {
          "description": "ProductID int64     `json:\"product_id\"`\nUnitID    int64 `json:\"unit_id\"`\nSupplierID int64    `json:\"supplier_id\"`\nSupplier   Supplier `pg:\"fk:supplier_id\" json:\"supplier\"`",
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/aiman-zaki/go_dz_http/models"
    },
...
    "User": {
      "description": "User represents the product for this application",
      "type": "object",
      "required": [
        "last_name"
      ],
      "properties": {
        "date_created": {
          "description": "the dateCreated for this user",
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateCreated"
        },
        "date_updated": {
          "description": "the dateUpdated for this user",
          "type": "string",
          "format": "date-time",
          "x-go-name": "DateUpdated"
        },
        "id": {
          "description": "the id for this user",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID",
          "readOnly": true
        },
        "last_name": {
          "description": "the last name for this user",
          "type": "string",
          "minLength": 3,
          "x-go-name": "LastName"
        },
        "role_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RoleID"
        },
        "salary": {
          "description": "the salary for this user",
          "type": "number",
          "format": "double",
          "x-go-name": "Salary"
        }
      },
      "x-go-package": "github.com/aiman-zaki/go_dz_http/models"
    }
  },
  "responses": {
...
    "notAuthorized": {
      "description": "NotAuthorized :",
      "schema": {
        "type": "object",
        "properties": {
          "Message": {
            "type": "string"
          }
        }
      }
    },
...
    "stock": {
      "description": "StockResponse :",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "stock": {
            "$ref": "#/definitions/Stock"
          }
        }
      }
    },
...
    "stocks": {
      "description": "StocksResponse :",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "stocks": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Stock"
            },
            "x-go-name": "Stock"
          }
        }
      }
    },
...
    "user": {
      "description": "UserResponse : List all users",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          },
          "user": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
    "users": {
      "description": "UsersResponse : List all users",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "the success message",
            "type": "string",
            "x-go-name": "Message"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/User"
            },
            "x-go-name": "User"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "flow": "implicit"
    }
  },
  "security": [
    {
      "Bearer": [
        "[]"
      ]
    }
  ]
}